requirements.txt
langchain
openai
huggingface_hub
python-dotenv
streamlit
langchain_community
unstructured
tiktoken
pinecone-client
pypdf
openai
langchain
pandas 
numpy
python-dotenv
langchain_openai
cassio
datasets
tiktoken
PyPDF2